requirements_with_gpu.txt
ta==0.4.7
stable-baselines==2.8.0
tensorflow-gpu==1.14.0

gym>=0.12.5
pandas>=0.23.4
matplotlib>=3.0.2
statsmodels>=0.10.2

numpy<=1.15.4

optuna
seaborn
tqdm
nltk
